requirements.txt
pip
setuptools
wheel
sqlalchemy
alembic
python-dotenv
gunicorn
celery
flower
fastapi[all]
fastapi-cache2[redis]
httpx
token-throttler